fix: change modules

+++ src/role/role.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateRoleDto } from './dto/create-role.dto';
 import { UpdateRoleDto } from './dto/update-role.dto';
 import { PrismaService } from 'prisma/prisma.service';
@@ -11,7 +15,10 @@ export class RoleService {
     return this.prisma.role.findUnique({ where: { name: 'User' } });
   }
 
-  async createRole(createRoleDto: CreateRoleDto) {
+  async createRole(createRoleDto: CreateRoleDto, token: string) {
+    if (token['roleId'] !== 1) {
+      throw new ForbiddenException('Access denied, user not Admin');
+    }
     return this.prisma.role.create({
       data: {
         name: createRoleDto.name,
@@ -20,11 +27,17 @@ export class RoleService {
     });
   }
 
-  async getAllRoles() {
+  async getAllRoles(token: string) {
+    if (token['roleId'] !== 1) {
+      throw new ForbiddenException('Access denied, user not Admin');
+    }
     return this.prisma.role.findMany();
   }
 
-  async create(createRoleDto: CreateRoleDto) {
+  async create(createRoleDto: CreateRoleDto, token: string) {
+    if (token['roleId'] !== 1) {
+      throw new ForbiddenException('Access denied, user not Admin');
+    }
     return this.prisma.role.create({
       data: {
         name: createRoleDto.name,
@@ -33,11 +46,10 @@ export class RoleService {
     });
   }
 
-  async findAll() {
-    return this.prisma.role.findMany();
-  }
-
-  async findOne(id: number) {
+  async findOne(id: number, token: string) {
+    if (token['roleId'] !== 1) {
+      throw new ForbiddenException('Access denied, user not Admin');
+    }
     const role = await this.prisma.role.findUnique({ where: { id } });
     if (!role) {
       throw new NotFoundException(`Role with ID ${id} not found`);
@@ -45,23 +57,11 @@ export class RoleService {
     return role;
   }
 
-  async updateUserRole(userId: number, roleName: string) {
-    const role = await this.prisma.role.findUnique({
-      where: { name: roleName },
-    });
-
-    if (!role) {
-      throw new NotFoundException(`Role with name ${roleName} not found`);
+  async update(id: number, updateRoleDto: UpdateRoleDto, token: string) {
+    if (token['roleId'] !== 1) {
+      throw new ForbiddenException('Access denied, user not Admin');
     }
-
-    return this.prisma.user.update({
-      where: { id: userId },
-      data: { roleId: role.id },
-    });
-  }
-
-  async update(id: number, updateRoleDto: UpdateRoleDto) {
-    const existingRole = await this.findOne(id);
+    const existingRole = await this.findOne(id, token);
     return this.prisma.role.update({
       where: { id: existingRole.id },
       data: {
@@ -71,8 +71,11 @@ export class RoleService {
     });
   }
 
-  async remove(id: number) {
-    const existingRole = await this.findOne(id);
+  async remove(id: number, token: string) {
+    if (token['roleId'] !== 1) {
+      throw new ForbiddenException('Access denied, user not Admin');
+    }
+    const existingRole = await this.findOne(id, token);
     return this.prisma.role.delete({ where: { id: existingRole.id } });
   }
 }
